+++ Changelog.md
@@ -4,6 +4,13 @@
    set the received message rate on the subscriber side.
     * [Pull request 174](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/174)
 
+1. Do not allow to advertise a topic that is currently advertised on the same node.
+   See [issue #54] (https://bitbucket.org/ignitionrobotics/ign-transport/issues/54)
+    * [Pull request 169](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/169)
+
+1. ZeroMQ updated from 3.2.4 to 4.0.4 on Windows.
+    * [Pull request 171](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/171)
+
 ## Ignition Transport 2.x
 
 1. Copyright added to `tools/code_check.sh` and `tools/cpplint_to_cppcheckxml.py`